% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_c.R
\name{f_to_c}
\alias{f_to_c}
\title{Fahrenheit Converter}
\usage{
f_to_c(temp_F)
}
\arguments{
\item{temp_F}{Numeric input indicating temperature in fahrenheit.}
}
\value{
Numeric representing \code{temp_F} converted to celsius.
}
\description{
This function converts from farenheit to celsius.
}
\examples{
f_to_c(32)

}
\keyword{temperature.}
