% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{Cross Validated K-Nearest Neighbors}
\usage{
my_knn_cv(data, cl, k_nn, k_cv)
}
\arguments{
\item{data}{Numeric input indicating temperature in fahrenheit.}

\item{cl}{true class of training data.}

\item{k_nn}{Integer for number of neighbors.}

\item{k_cv}{Integer for number of folds in cross validation.}
}
\value{
List of 2 outputs: Vector of predicted class for all observations. Numeric for cross validated misclassification error
}
\description{
This function predicts class based off of covariates using KNN methods. Note to remove NA
}
\examples{
penguins2 <- na.omit(my_penguins)
my_knn_cv(penguins2[c("flipper_length_mm", "bill_depth_mm")], penguins2["species"], 5, 5)
}
\keyword{learning}
\keyword{machine}
