% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object indicating your formula.}

\item{data}{Data to fit to model.}
}
\value{
Returns a table which tells coeffecient estimate, standard error, t value, and p value for each coeffecient as well as intercept.
}
\description{
This function creates a linear model for your data.
}
\examples{
my_lm(bill_length_mm~bill_depth_mm, my_penguins)


}
\keyword{regression}
