% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_penguins.R
\docType{data}
\name{my_penguins}
\alias{my_penguins}
\title{Penguins data.}
\format{
The main data frame \code{palmerpenguins} has 344 rows and 8 variables:
\describe{
\item{species}{Factor of 3 levels describing what species}
\item{island}{Factor of 3 levels describing what island}
\item{bill_length_mm}{Lengthof bill in millimeters}
\item{bill_depth_mm}{Depth of bill in millimeters}
\item{body_mass_g}{Mass of Penguin in grams}
\item{sex}{Factor of 2 levels describing sex of Penguin}
\item{year}{Year observed, 2007-2009}
}
}
\usage{
my_penguins
}
\description{
This is some descriptive data collected on penguins
This data and documentation come from the \code{palmerpenguins} package,
available as \code{\link[palmerpenguins]{palmerpenguins}}.
}
\examples{
penguins <- my_penguins
str(penguins)
head(penguins)
summary(penguins)
table(penguins$island)

}
\keyword{datasets}
