% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{Specifiys nature of alternative hypothesis. \code{two.sided}, \code{less}, \code{greater}.}

\item{mu}{Numeric indicating null hypothesis value of the mean.}
}
\value{
Returns a list with 4 peices of info: Numeric test statistic, Numeric degrees of freedom, String stating the method used, Numeric P value.
}
\description{
This function performs a one sample T test.
}
\examples{
x <- c(34,32,36,2,34,33,29,21,24)
mu <- 30
my_t.test(x,"two.sided", mu)

my_t.test(na.omit(my_penguins$bill_depth_mm), "less", 40)

}
\keyword{statistical}
\keyword{test}
